Code Cleanup Modified fix for null errors
