+++ VotingPlugin/src/com/Ben12345rocks/VotingPlugin/Config/Config.java
@@ -44,7 +44,7 @@ public class Config {
 				plugin.saveResource("Config.yml", true);
 			} catch (IOException e) {
 				Bukkit.getServer().getLogger()
-						.severe(ChatColor.RED + "Could not create Config.yml!");
+				.severe(ChatColor.RED + "Could not create Config.yml!");
 			}
 		}
 
@@ -60,7 +60,7 @@ public class Config {
 			data.save(dFile);
 		} catch (IOException e) {
 			Bukkit.getServer().getLogger()
-					.severe(ChatColor.RED + "Could not save Config.yml!");
+			.severe(ChatColor.RED + "Could not save Config.yml!");
 		}
 	}
