+++ VotingPlugin/src/com/Ben12345rocks/VotingPlugin/Config/ConfigBonusReward.java
@@ -47,9 +47,9 @@ public class ConfigBonusReward {
 				plugin.saveResource("BonusReward.yml", true);
 			} catch (IOException e) {
 				Bukkit.getServer()
-						.getLogger()
-						.severe(ChatColor.RED
-								+ "Could not create BonusReward.yml!");
+				.getLogger()
+				.severe(ChatColor.RED
+						+ "Could not create BonusReward.yml!");
 			}
 		}
 
@@ -65,7 +65,7 @@ public class ConfigBonusReward {
 			data.save(dFile);
 		} catch (IOException e) {
 			Bukkit.getServer().getLogger()
-					.severe(ChatColor.RED + "Could not save BonusReward.yml!");
+			.severe(ChatColor.RED + "Could not save BonusReward.yml!");
 		}
 	}
