+++ VotingPlugin/src/com/Ben12345rocks/VotingPlugin/Events/PlayerJoinEvent.java
@@ -40,24 +40,24 @@ public class PlayerJoinEvent implements Listener {
 		userDataFile(user);
 
 		UUIDs.getInstance()
-				.setName(playerName, player.getUniqueId().toString());
+		.setName(playerName, player.getUniqueId().toString());
 
 		plugin.getServer().getScheduler()
-				.runTaskLaterAsynchronously(plugin, new Runnable() {
-					@Override
-					public void run() {
-						// msg player if there is a update
-						if (Config.getInstance().updateReminder()) {
-							updateCheckLogin(player);
-						}
+		.runTaskLaterAsynchronously(plugin, new Runnable() {
+			@Override
+			public void run() {
+				// msg player if there is a update
+				if (Config.getInstance().updateReminder()) {
+					updateCheckLogin(player);
+				}
 
-						// give offline vote (if they voted offline)
-						user.offVote();
+				// give offline vote (if they voted offline)
+				user.offVote();
 
-						// msg player if he can vote
-						user.loginMessage();
-					}
-				}, 100L);
+				// msg player if he can vote
+				user.loginMessage();
+			}
+		}, 100L);
 	}
 
 	private void updateCheckLogin(Player player) {
@@ -74,9 +74,9 @@ public class PlayerJoinEvent implements Listener {
 			if (player.hasPermission("VotingPlugin.remindupdate")) {
 				player.sendMessage(Utils.getInstance().colorize(
 						"&a" + plugin.getName()
-								+ " has an update available! Your Version: &c"
-								+ plugin.getDescription().getVersion()
-								+ "&a New Version: &c" + updater.getVersion()));
+						+ " has an update available! Your Version: &c"
+						+ plugin.getDescription().getVersion()
+						+ "&a New Version: &c" + updater.getVersion()));
 			}
 			break;
 		}
