+++ VotingPlugin/src/com/Ben12345rocks/VotingPlugin/Metrics/Metrics.java
@@ -131,7 +131,7 @@ public class Metrics {
 		// Do we need to create the file?
 		if (configuration.get("guid", null) == null) {
 			configuration.options().header("http://mcstats.org")
-					.copyDefaults(true);
+			.copyDefaults(true);
 			configuration.save(configurationFile);
 		}
 
@@ -215,9 +215,9 @@ public class Metrics {
 								synchronized (optOutLock) {
 									// Disable Task, if it is running and the
 									// server owner decided to opt-out
-									if (isOptOut() && taskId > 0) {
+									if (isOptOut() && (taskId > 0)) {
 										plugin.getServer().getScheduler()
-												.cancelTask(taskId);
+										.cancelTask(taskId);
 										taskId = -1;
 									}
 								}
@@ -402,7 +402,7 @@ public class Metrics {
 		writer.close();
 		reader.close();
 
-		if (response == null || response.startsWith("ERR")) {
+		if ((response == null) || response.startsWith("ERR")) {
 			throw new IOException(response); // Throw the exception
 		} else {
 			// Is this the first update this hour?
@@ -455,9 +455,9 @@ public class Metrics {
 	 */
 	private static void encodeDataPair(final StringBuilder buffer,
 			final String key, final String value)
-			throws UnsupportedEncodingException {
+					throws UnsupportedEncodingException {
 		buffer.append('&').append(encode(key)).append('=')
-				.append(encode(value));
+		.append(encode(value));
 	}
 
 	/**
@@ -605,7 +605,7 @@ public class Metrics {
 
 			final Plotter plotter = (Plotter) object;
 			return plotter.name.equals(name)
-					&& plotter.getValue() == getValue();
+					&& (plotter.getValue() == getValue());
 		}
 
 	}
