+++ VotingPlugin/src/com/Ben12345rocks/VotingPlugin/Updater/Updater.java
@@ -43,7 +43,7 @@ public class Updater {
 
 		try {
 			connection = (HttpURLConnection) new URL(HOST + QUERY)
-					.openConnection();
+			.openConnection();
 		} catch (IOException e) {
 			result = UpdateResult.FAIL_SPIGOT;
 			return;
