+++ VotingPlugin/src/com/Ben12345rocks/VotingPlugin/UserData/UUIDs.java
@@ -43,7 +43,7 @@ public class UUIDs {
 				dFile.createNewFile();
 			} catch (IOException e) {
 				Bukkit.getServer().getLogger()
-						.severe(ChatColor.RED + "Could not create uuids.yml!");
+				.severe(ChatColor.RED + "Could not create uuids.yml!");
 			}
 		}
 
@@ -59,7 +59,7 @@ public class UUIDs {
 			data.save(dFile);
 		} catch (IOException e) {
 			Bukkit.getServer().getLogger()
-					.severe(ChatColor.RED + "Could not save uuids.yml!");
+			.severe(ChatColor.RED + "Could not save uuids.yml!");
 		}
 	}
